% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistance.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Calculate the Genetic Distance Between Samples}
\usage{
calculateDistance(snp_matrix)
}
\arguments{
\item{snp_matrix}{Matrix containing SNP calls
Matrix containing genetic distance between all samples}
}
\description{
Calculates the percentage of SNP matches. The number of complete and matching
SNP loci is the numerator and the number of SNPs non-missing for both samples
is the denominator
}
