% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSnps.R
\name{getSnps}
\alias{getSnps}
\title{Get Table of SNPs by Sample}
\usage{
getSnps(subsetByChromosome = NULL, snps = snps, input = input)
}
\arguments{
\item{subsetByChromosome}{Chromosome number to use for comparison, left NULL if using SNP list}

\item{snps}{List of SNP IDs to use for comparison}

\item{input}{Input Directory Containing annotated VCF files}
}
\value{
Matrix containing SNP calls for all samples
}
\description{
This funtion reads a set of VCF files and creates a table of SNPs with SNP IDs as rows and
sample names (one per file) as columns
}
