% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSniffer.R
\name{sampleSniffer}
\alias{sampleSniffer}
\title{Finds the Closest Genetic Match for Each Sample}
\usage{
sampleSniffer(input = NULL, snps = NULL, chromosome = NULL,
  output = "sampleSniffer.pdf")
}
\arguments{
\item{input}{Input directory containing VCF files}

\item{snps}{Name of file containing the SNP IDs to be used}

\item{output}{Name of output file}

\item{chromsome}{Chromosome number to be used}
}
\description{
Reads in all VCF files in input directory, calculates genetic distance between each
pair of samples, and finds the closest genetic match for each sample. Either snps
file or chromosome number must be specified, but only one can be used at a time.
Valid values for chromosome option are numbers 1-23 and X. If output filename is
given, value is defaulted to sampleSniffer.pdf.
}
